getSass should always return a Promise
Fixes #3

In order to get a Sass object that can be instantiated we first have to
look up whether the web worker is located. To do this the `getSass`
function was created. After it has done the work the first time getSass
simply returns the same sass object. However the use of getSass expects
it to return a Promise, which means it will start throwing after the 2nd
scss file is imported. The fix is simply to always resolve getSass with
a Promise.resolve call.